% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_modules.R
\name{install_modules}
\alias{install_modules}
\title{Downloads and install modules for python}
\usage{
install_modules(modules, upgrade_pip = T)
}
\arguments{
\item{modules}{Specify a character vector of all python modules to install}

\item{upgrade_pip}{Specify whether to upgrade the module manager before install i.e. to get the most updated versions for your python version}
}
\value{
Installed python modules
}
\description{
Runs command line installation of python modules and can select whether to upgrade the python module manager 'pip'.
}
\examples{
install_modules(modules = c("keras", "tensorflow", "scikit-learn==0.19.1"))
install_modules(modules = c("keras", "tensorflow", "scikit-learn==0.19.1"), upgrade_pip = F)
}
