% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_python.R
\name{link_python}
\alias{link_python}
\title{Links python to R}
\usage{
link_python()
}
\value{
Linked python program
}
\description{
To run python through R, its necassary ot link python to R. link_python() recommends a potential python location and prompts the user to select it.
}
\examples{
link_python()
}
