% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_download.R
\name{py_download}
\alias{py_download}
\title{Downloads and prepares the install of python}
\usage{
py_download(version = "3.6.8")
}
\arguments{
\item{version}{Specify a python version as a character string. Refer to "https://www.python.org/downloads/" to see available versions}
}
\value{
An installed version of python
}
\description{
This function downloads and prepares the install of python. Users can select the specific version of python, with the default set at 3.6.8 - a stable version to utilise python's machine learning languages. When prompted to install, you are given the choice to add Python to the PATH. If you only plan to use one version of python, this would be adviseable. This function currently only works for Windows operating systems.
}
\details{
For the python install to be recognised by R, its important to close and re-open R, which resets the PATH's. As a result, we recommend saving scripts and important files before running this function. R will only close if confirmed by you (the user).
}
\examples{
py_download()
py_download(version = "3.9.0")
}
