% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_python.R
\name{link_python}
\alias{link_python}
\title{Links python to R}
\usage{
link_python(set_path = T)
}
\arguments{
\item{set_path}{Specify whether you want to manually link R and python. Binary: TRUE/FALSE}
}
\value{
Linked python program
}
\description{
To run python through R, its necassary ot link python to R. This can be done manually by opting to specify the python.exe file path, or you can allow the function search for a python.exe file.
}
\examples{
link_python(set_path = T)
}
